# backend/internal/api/specs.go
// Package api provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/deepmap/oapi-codegen version v1.16.3 DO NOT EDIT.
package api

// Defines values for ClueSetAxis.
const (
	X ClueSetAxis = "X"
	Y ClueSetAxis = "Y"
	Z ClueSetAxis = "Z"
)

// Defines values for ClueSplit.
const (
	NoSplit    ClueSplit = "NoSplit"
	Split2     ClueSplit = "Split2"
	Split3Plus ClueSplit = "Split3Plus"
)

// ActionResult defines model for ActionResult.
type ActionResult struct {
	Completed      bool `json:"completed"`
	CompletionTime *int `json:"completionTime,omitempty"`
	RemainingLives int  `json:"remainingLives"`
	Success        bool `json:"success"`
}

// ClueSet defines model for ClueSet.
type ClueSet struct {
	Axis   ClueSetAxis `json:"axis"`
	Coord1 int         `json:"coord1"`
	Coord2 int         `json:"coord2"`
	Count  *int        `json:"count"`
	Split  ClueSplit   `json:"split"`
}

// ClueSetAxis defines model for ClueSet.Axis.
type ClueSetAxis string

// ClueSplit defines model for ClueSplit.
type ClueSplit string

// Puzzle defines model for Puzzle.
type Puzzle struct {
	Author string    `json:"author"`
	Clues  []ClueSet `json:"clues"`
	Id     string    `json:"id"`
	Name   string    `json:"name"`
	SizeX  int       `json:"sizeX"`
	SizeY  int       `json:"sizeY"`
	SizeZ  int       `json:"sizeZ"`
}

// PuzzleSummary defines model for PuzzleSummary.
type PuzzleSummary struct {
	Author string `json:"author"`
	Id     string `json:"id"`
	Name   string `json:"name"`
}

// VoxelAction defines model for VoxelAction.
type VoxelAction struct {
	X int `json:"x"`
	Y int `json:"y"`
	Z int `json:"z"`
}

// PostApiPuzzlesIdActionsJSONRequestBody defines body for PostApiPuzzlesIdActions for application/json ContentType.
type PostApiPuzzlesIdActionsJSONRequestBody = VoxelAction